Gtkdatabox is no longer mandatory and enable Gtk-3
If you don't have gtkdatabox installed, you can build the
capture and the plethysmograph programs, oscilloscope and
file viewer won't be built, because gtkdatabox is required
for these programs.

gtk3 detection in the configure script it's enabled again
but note that the code it's still dependent on gtk2, so
the build will fail if you have both gtk3 and gtk2 installed.

Signed-off-by: Víctor Enríquez <victor.quicksilver@gmail.com>